+++ idea/testData/hierarchy/calls/callersJava/javaMethod/main1.kt
@@ -1,0 +1,11 @@
+class A {
+    val x = J().foo()
+
+    init {
+        J().foo()
+    }
+}
+
+val y = J().foo()
+
+fun test(j: J) = j.foo()